+++ src/components/Notes/Note.js
